requirements.txt
git+https://github.com/VRSEN/agency-swarm.git@feature/multimodal-tool-outputs
openai-agents[litellm]
beautifulsoup4>=4.9.0
dulwich>=0.21.6
html2text>=2020.1.16
jupyter>=1.0.0
markdown>=3.3.0
nbformat>=5.0.0
pydantic>=2.0.0
pytest>=6.0.0
pytest-asyncio>=0.21.0
requests>=2.25.0
google-genai>=0.3.0
pillow>=9.0.0
selenium>=4.0.0
webdriver-manager>=3.8.0
matplotlib>=3.5.0